+++ src/main/java/com/foodDelivery/project/repository/OrderRepository.java
@@ -11,8 +11,8 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 
     @Query(value = "SELECT * FROM orders WHERE user_id = :id ORDER BY createdAt DESC", nativeQuery = true)
-    List<Order> getOrdersByUserIdOrderByCreatedAtDesc(@Param("id") int id);
+    List<Order> getOrdersByUserIdOrderByCreatedAtDesc(@Param("id") Long id);
 }
